+++ src/main/java/com/metDaisy/clickerheroes/mapper/HeroSkillMapper.java
@@ -2,6 +2,7 @@ package com.metDaisy.clickerheroes.mapper;
 
 import com.metDaisy.clickerheroes.dto.HeroSkillDto;
 import com.metDaisy.clickerheroes.entity.HeroSkill;
+import com.metDaisy.clickerheroes.mapper.base.BaseMapper;
 import com.metDaisy.clickerheroes.mapper.config.GlobalMapperConfig;
 import org.mapstruct.Mapper;
 import org.mapstruct.MappingTarget;
@@ -10,8 +11,6 @@ import org.mapstruct.MappingTarget;
     config = GlobalMapperConfig.class,
     uses = {HeroMapper.class, ScientificNumberMapper.class})
 public interface HeroSkillMapper extends BaseMapper<HeroSkill, HeroSkillDto> {
-
-  HeroSkill toEntity(HeroSkillDto heroSkillDto);
 
   HeroSkill partialUpdate(HeroSkillDto heroSkillDto, @MappingTarget HeroSkill heroSkill);
 }
