+++ src/main/java/com/metDaisy/clickerheroes/mapper/HiredHeroMapper.java
@@ -2,6 +2,7 @@ package com.metDaisy.clickerheroes.mapper;
 
 import com.metDaisy.clickerheroes.dto.HiredHeroDto;
 import com.metDaisy.clickerheroes.entity.HiredHero;
+import com.metDaisy.clickerheroes.mapper.base.BaseMapper;
 import com.metDaisy.clickerheroes.mapper.config.GlobalMapperConfig;
 import org.mapstruct.Mapper;
 import org.mapstruct.MappingTarget;
@@ -10,8 +11,6 @@ import org.mapstruct.MappingTarget;
     config = GlobalMapperConfig.class,
     uses = {UserMapper.class, HeroMapper.class, ScientificNumberMapper.class})
 public interface HiredHeroMapper extends BaseMapper<HiredHero, HiredHeroDto> {
-
-  HiredHero toEntity(HiredHeroDto hiredHeroDto);
 
   HiredHero partialUpdate(HiredHeroDto hiredHeroDto, @MappingTarget HiredHero hiredHero);
 }
