+++ src/main/java/com/metDaisy/clickerheroes/mapper/MonsterMapper.java
@@ -2,6 +2,7 @@ package com.metDaisy.clickerheroes.mapper;
 
 import com.metDaisy.clickerheroes.dto.MonsterDto;
 import com.metDaisy.clickerheroes.entity.Monster;
+import com.metDaisy.clickerheroes.mapper.base.BaseMapper;
 import com.metDaisy.clickerheroes.mapper.config.GlobalMapperConfig;
 import org.mapstruct.Mapper;
 import org.mapstruct.MappingTarget;
@@ -10,8 +11,6 @@ import org.mapstruct.MappingTarget;
     config = GlobalMapperConfig.class,
     uses = {UserMapper.class, ScientificNumberMapper.class})
 public interface MonsterMapper extends BaseMapper<Monster, MonsterDto> {
-
-  Monster toEntity(MonsterDto monsterDto);
 
   Monster partialUpdate(MonsterDto monsterDto, @MappingTarget Monster monster);
 }
