+++ src/main/java/com/metDaisy/clickerheroes/mapper/UserMapper.java
@@ -3,6 +3,7 @@ package com.metDaisy.clickerheroes.mapper;
 import com.metDaisy.clickerheroes.dto.UserDto;
 import com.metDaisy.clickerheroes.dto.request.UserCreateRequest;
 import com.metDaisy.clickerheroes.entity.User;
+import com.metDaisy.clickerheroes.mapper.base.BaseMapper;
 import com.metDaisy.clickerheroes.mapper.config.GlobalMapperConfig;
 import org.mapstruct.AfterMapping;
 import org.mapstruct.Mapper;
